+++ src/main/java/com/newjob/consultant/controller/admin/dto/AdminForm.java
@@ -1,4 +1,4 @@
-package com.newjob.consultant.entity.consultant.dto;
+package com.newjob.consultant.controller.admin.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -8,8 +8,7 @@ import lombok.Setter;
 @Getter @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-public class ConsultantLoginForm {
-    private Long id;
-    private String email;
+public class AdminForm {
+    private String id;
     private String password;
 }
